This should now compile again with the latest FIFE trunk. Note only release build works still. I believe it has something to do with incompatible DLLs being loaded for SDL (debug vs. release). I cannot seem to force the debug dlls to be loaded since not all dependent DLLs were built against the debug version of SDL.

+++ src/tutorial_1/ScreenScroller.cpp
@@ -35,10 +35,10 @@ ScreenScroller::ScreenScroller(FIFE::Camera* camera, FIFE::EventManager* eventMa
 		const FIFE::Rect& viewport = m_camera->getViewPort();
 
 		// calculate borders to activate automatic scrolling
-		m_scrollAreaTop = viewport.h - (viewport.h*ScrollActivationPercent);
-		m_scrollAreaBottom = viewport.y + (viewport.h*ScrollActivationPercent);
-		m_scrollAreaRight = viewport.w - (viewport.w*ScrollActivationPercent);
-		m_scrollAreaLeft = viewport.x + (viewport.w*ScrollActivationPercent);
+        m_scrollAreaTop = static_cast<int>(viewport.h - (viewport.h*ScrollActivationPercent));
+		m_scrollAreaBottom = static_cast<int>(viewport.y + (viewport.h*ScrollActivationPercent));
+		m_scrollAreaRight = static_cast<int>(viewport.w - (viewport.w*ScrollActivationPercent));
+		m_scrollAreaLeft = static_cast<int>(viewport.x + (viewport.w*ScrollActivationPercent));
 	}
 }
